move tests into dataclasses_config.py
